Task 1 of 10-June-2024 completed

+++ Date_10_June_2024_Home_Task/Task002.py
@@ -13,5 +13,5 @@ print("Cube of a number is ",  math.pow(number, 3))
 # solution for second problem of this task
 # a = int(input("Enter value for variable a: \t"))
 # b = int(input("Enter value for variable b: \t"))
-a, b = int(input("Enter value for variable a : \n")), int(input("Enter value for variable b : \n"))
+a, b = float(input("Enter value for variable a : \n")), float(input("Enter value for variable b : \n"))
 print("a is greater than b" if a > b else "a is lesser than b" if a < b else "a is equal to b")
